requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
langchain-experimental
langgraph
python-dotenv
pydantic
httpx
requests
